Feat: create api default header [#11]
