+++ remote/src/main/java/retrofit/ApiService.kt
@@ -1,0 +1,22 @@
+package retrofit
+
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object ApiService {
+
+    private const val BASE_URL = "https://api.upbit.com/v1" // TODO: v1을 빼야할수도
+
+    val upbitClient by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .client(provideOkHttpClient(AppInterceptor()))
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+
+    private fun provideOkHttpClient(interceptor: AppInterceptor) = OkHttpClient.Builder()
+        .addInterceptor(interceptor)
+        .build()
+}
